requirements.txt
flask == 3.1.1
flask-cors == 6.0.0
google-generativeai == 0.8.5
python-dotenv == 1.1.0
PyPDF2 == 3.0.1
gunicorn